split ProviderMap away from QueryHandler; fixes #938
