+++ koku/api/report/queries.py
@@ -24,19 +24,11 @@ from decimal import Decimal, DivisionByZero, InvalidOperation
 from itertools import groupby
 from urllib.parse import quote_plus
 
-from django.db.models import CharField, Count, DecimalField, F, Max, Q, Sum, Value
-from django.db.models.expressions import ExpressionWrapper, OrderBy, RawSQL
-from django.db.models.functions import Coalesce
-from providers.provider_access import ProviderAccessor
+from django.db.models import Q
+from django.db.models.expressions import OrderBy, RawSQL
 
 from api.query_filter import QueryFilter, QueryFilterCollection
 from api.query_handler import QueryHandler
-from reporting.models import (AWSCostEntryLineItemDailySummary,
-                              CostSummary,
-                              OCPAWSCostLineItemDailySummary,
-                              OCPAWSCostLineItemProjectDailySummary,
-                              OCPStorageLineItemDailySummary,
-                              OCPUsageLineItemDailySummary)
 
 LOG = logging.getLogger(__name__)
 
@@ -46,760 +38,10 @@ def strip_tag_prefix(tag):
     return tag.replace('tag:', '').replace('and:', '').replace('or:', '')
 
 
-class ProviderMap(object):
-    """Data structure mapping between API params and DB Model names.
-
-    The idea here is that reports ought to be operating on largely similar
-    data - counts, costs, etc. The only variable is determining which
-    DB tables and fields are supplying the requested data.
-
-    ProviderMap supplies ReportQueryHandler with the appropriate model
-    references.
-    """
-
-    PACK_DEFINITIONS = {
-        'cost': {
-            'keys': ['cost', 'infrastructure_cost', 'derived_cost'],
-            'units': 'cost_units'
-        },
-        'usage': {
-            'keys': ['usage', 'request', 'limit', 'capacity'],
-            'units': 'usage_units'
-        },
-        'count': {
-            'keys': ['count'],
-            'units': 'count_units'
-        }
-    }
-
-    # main mapping data structure
-    # this data should be considered static and read-only.
-    MAPPING = [
-        {
-            'provider': 'AWS',
-            'alias': 'account_alias__account_alias',
-            'annotations': {
-                'account': 'usage_account_id',
-                'service': 'product_code',
-                'az': 'availability_zone'
-            },
-            'end_date': 'usage_end',
-            'filters': {
-                'account': [
-                    {
-                        'field': 'account_alias__account_alias',
-                        'operation': 'icontains',
-                        'composition_key': 'account_filter'
-                    },
-                    {
-                        'field': 'usage_account_id',
-                        'operation': 'icontains',
-                        'composition_key': 'account_filter'
-                    }
-                ],
-                'service': {
-                    'field': 'product_code',
-                    'operation': 'icontains'
-                },
-                'az': {
-                    'field': 'availability_zone',
-                    'operation': 'icontains'
-                },
-                'region': {
-                    'field': 'region',
-                    'operation': 'icontains'
-                },
-                'product_family': {
-                    'field': 'product_family',
-                    'operation': 'icontains'
-                }
-            },
-            'group_by_options': ['service', 'account', 'region', 'az', 'product_family'],
-            'tag_column': 'tags',
-            'report_type': {
-                'costs': {
-                    'aggregates': {
-                        'infrastructure_cost': Sum('unblended_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('unblended_cost')
-                    },
-                    'aggregate_key': 'unblended_cost',
-                    'annotations': {
-                        'infrastructure_cost': Sum('unblended_cost'),
-                        'derived_cost': Value(0, output_field=DecimalField()),
-                        'cost': Sum('unblended_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD'))
-                    },
-                    'delta_key': {'cost': Sum('unblended_cost')},
-                    'filter': {},
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'sum_columns': ['cost', 'infrastructure_cost', 'derived_cost'],
-                    'default_ordering': {'cost': 'desc'},
-                },
-                'instance_type': {
-                    'aggregates': {
-                        'infrastructure_cost': Sum('unblended_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('unblended_cost'),
-                        'count': Sum(Value(0, output_field=DecimalField())),
-                        'usage': Sum('usage_amount'),
-                    },
-                    'aggregate_key': 'usage_amount',
-                    'annotations': {
-                        'infrastructure_cost': Sum('unblended_cost'),
-                        'derived_cost': Value(0, output_field=DecimalField()),
-                        'cost': Sum('unblended_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        # The summary table already already has counts
-                        'count': Max('resource_count'),
-                        'count_units': Value('instances', output_field=CharField()),
-                        'usage': Sum('usage_amount'),
-                        'usage_units': Coalesce(Max('unit'), Value('Hrs'))
-                    },
-                    'delta_key': {'usage': Sum('usage_amount')},
-                    'filter': {
-                        'field': 'instance_type',
-                        'operation': 'isnull',
-                        'parameter': False
-                    },
-                    'group_by': ['instance_type'],
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'usage_units_key': 'unit',
-                    'usage_units_fallback': 'Hrs',
-                    'count_units_fallback': 'instances',
-                    'sum_columns': ['usage', 'cost', 'infrastructure_cost', 'derived_cost', 'count'],
-                    'default_ordering': {'usage': 'desc'},
-                },
-                'storage': {
-                    'aggregates': {
-                        'usage': Sum('usage_amount'),
-                        'infrastructure_cost': Sum('unblended_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('unblended_cost')
-                    },
-                    'aggregate_key': 'usage_amount',
-                    'annotations': {
-                        'infrastructure_cost': Sum('unblended_cost'),
-                        'derived_cost': Value(0, output_field=DecimalField()),
-                        'cost': Sum('unblended_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'usage': Sum('usage_amount'),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'delta_key': {'usage': Sum('usage_amount')},
-                    'filter': {
-                        'field': 'product_family',
-                        'operation': 'contains',
-                        'parameter': 'Storage'
-                    },
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'usage_units_key': 'unit',
-                    'usage_units_fallback': 'GB-Mo',
-                    'sum_columns': ['usage', 'cost', 'infrastructure_cost', 'derived_cost'],
-                    'default_ordering': {'usage': 'desc'},
-                },
-            },
-            'start_date': 'usage_start',
-            'tables': {
-                'query': AWSCostEntryLineItemDailySummary,
-            },
-        },
-        {
-            'provider': 'OCP',
-            'annotations': {
-                'cluster': 'cluster_id',
-                'project': 'namespace'
-            },
-            'end_date': 'usage_end',
-            'filters': {
-                'project': {
-                    'field': 'namespace',
-                    'operation': 'icontains'
-                },
-                'cluster': [
-                    {
-                        'field': 'cluster_alias',
-                        'operation': 'icontains',
-                        'composition_key': 'cluster_filter'
-                    },
-                    {
-                        'field': 'cluster_id',
-                        'operation': 'icontains',
-                        'composition_key': 'cluster_filter'
-                    }
-                ],
-                'pod': {
-                    'field': 'pod',
-                    'operation': 'icontains'
-                },
-                'node': {
-                    'field': 'node',
-                    'operation': 'icontains'
-                },
-                'infrastructures': {
-                    'field': 'cluster_id',
-                    'operation': 'exact',
-                    'custom': ProviderAccessor('OCP').infrastructure_key_list
-                },
-            },
-            'group_by_options': ['cluster', 'project', 'node'],
-            'tag_column': 'pod_labels',
-            'report_type': {
-                'costs': {
-                    'tables': {
-                        'query': CostSummary
-                    },
-                    'aggregates': {
-                        'infrastructure_cost': Sum(F('infra_cost')),
-                        'derived_cost': Sum(
-                            ExpressionWrapper(
-                                F('pod_charge_cpu_core_hours') + \
-                                F('pod_charge_memory_gigabyte_hours') + \
-                                F('persistentvolumeclaim_charge_gb_month'),
-                                output_field=DecimalField()
-                            )
-                        ),
-                        'cost': Sum(
-                            ExpressionWrapper(
-                                F('pod_charge_cpu_core_hours') + \
-                                F('pod_charge_memory_gigabyte_hours') + \
-                                F('persistentvolumeclaim_charge_gb_month') + \
-                                F('infra_cost'),
-                                output_field=DecimalField()
-                            )
-
-                        ),
-                    },
-                    'default_ordering': {'cost': 'desc'},
-                    'annotations': {
-                        'infrastructure_cost': Sum(F('infra_cost')),
-                        'derived_cost': Sum(
-                            ExpressionWrapper(
-                                F('pod_charge_cpu_core_hours') + \
-                                F('pod_charge_memory_gigabyte_hours') + \
-                                F('persistentvolumeclaim_charge_gb_month'),
-                                output_field=DecimalField()
-                            )
-                        ),
-                        'cost': Sum(
-                            ExpressionWrapper(
-                                F('pod_charge_cpu_core_hours') + \
-                                F('pod_charge_memory_gigabyte_hours') + \
-                                F('persistentvolumeclaim_charge_gb_month') + \
-                                F('infra_cost'),
-                                output_field=DecimalField()
-                            )
-                        ),
-                        'cost_units': Value('USD', output_field=CharField())
-                    },
-                    'capacity_aggregate': {},
-                    'delta_key': {
-                        'cost': Sum(
-                            F('pod_charge_cpu_core_hours') +  # noqa: W504
-                            F('pod_charge_memory_gigabyte_hours') +  # noqa: W504
-                            F('persistentvolumeclaim_charge_gb_month') +  # noqa: W504
-                            F('infra_cost')
-                        )
-                    },
-                    'filter': {},
-                    'cost_units_key': 'USD',
-                    'sum_columns': ['cost', 'infrastructure_cost', 'derived_cost'],
-                },
-                'costs_by_project': {
-                    'tables': {
-                        'query': CostSummary
-                    },
-                    'aggregates': {
-                        'infrastructure_cost': Sum(F('project_infra_cost')),
-                        'derived_cost': Sum(F('pod_charge_cpu_core_hours') + \
-                                            F('pod_charge_memory_gigabyte_hours')),
-                        'cost': Sum(F('pod_charge_cpu_core_hours') + \
-                                    F('pod_charge_memory_gigabyte_hours') + \
-                                    F('persistentvolumeclaim_charge_gb_month') + \
-                                    F('project_infra_cost')),
-                    },
-                    'default_ordering': {'cost': 'desc'},
-                    'annotations': {
-                        'infrastructure_cost': Sum(F('project_infra_cost')),
-                        'derived_cost': Sum(F('pod_charge_cpu_core_hours') + \
-                                            F('pod_charge_memory_gigabyte_hours')),
-                        'cost': Sum(F('pod_charge_cpu_core_hours') + \
-                                    F('pod_charge_memory_gigabyte_hours') + \
-                                    F('persistentvolumeclaim_charge_gb_month') + \
-                                    F('project_infra_cost')),
-                        'cost_units': Value('USD', output_field=CharField())
-                    },
-                    'capacity_aggregate': {},
-                    'delta_key': {
-                        'cost': Sum(
-                            F('pod_charge_cpu_core_hours') +  # noqa: W504
-                            F('pod_charge_memory_gigabyte_hours') +  # noqa: W504
-                            F('persistentvolumeclaim_charge_gb_month') +  # noqa: W504
-                            F('project_infra_cost')
-                        )
-                    },
-                    'filter': {},
-                    'cost_units_key': 'USD',
-                    'sum_columns': ['cost', 'infrastructure_cost', 'derived_cost'],
-                },
-                'cpu': {
-                    'aggregates': {
-                        'usage': Sum('pod_usage_cpu_core_hours'),
-                        'request': Sum('pod_request_cpu_core_hours'),
-                        'limit': Sum('pod_limit_cpu_core_hours'),
-                        'infrastructure_cost': Sum(Value(0, output_field=DecimalField())),
-                        'derived_cost': Sum('pod_charge_cpu_core_hours'),
-                        'cost': Sum('pod_charge_cpu_core_hours')
-                    },
-                    'capacity_aggregate': {
-                        'capacity': Max('cluster_capacity_cpu_core_hours')
-                    },
-                    'default_ordering': {'usage': 'desc'},
-                    'annotations': {
-                        'usage': Sum('pod_usage_cpu_core_hours'),
-                        'request': Sum('pod_request_cpu_core_hours'),
-                        'limit': Sum('pod_limit_cpu_core_hours'),
-                        'capacity': {
-                            'total': Max('total_capacity_cpu_core_hours'),
-                            'cluster': Max('cluster_capacity_cpu_core_hours'),
-                        },
-                        'infrastructure_cost': Value(0, output_field=DecimalField()),
-                        'derived_cost': Sum('pod_charge_cpu_core_hours'),
-                        'cost': Sum('pod_charge_cpu_core_hours'),
-                        'cost_units': Value('USD', output_field=CharField()),
-                        'usage_units': Value('Core-Hours', output_field=CharField())
-                    },
-                    'delta_key': {
-                        'usage': Sum('pod_usage_cpu_core_hours'),
-                        'request': Sum('pod_request_cpu_core_hours'),
-                        'cost': Sum('pod_charge_cpu_core_hours')
-                    },
-                    'filter': {},
-                    'cost_units_key': 'USD',
-                    'usage_units_key': 'Core-Hours',
-                    'sum_columns': ['usage', 'request', 'limit', 'infrastructure_cost',
-                                    'derived_cost', 'cost'],
-                },
-                'memory': {
-                    'aggregates': {
-                        'usage': Sum('pod_usage_memory_gigabyte_hours'),
-                        'request': Sum('pod_request_memory_gigabyte_hours'),
-                        'limit': Sum('pod_limit_memory_gigabyte_hours'),
-                        'infrastructure_cost': Sum(Value(0, output_field=DecimalField())),
-                        'derived_cost': Sum('pod_charge_memory_gigabyte_hours'),
-                        'cost': Sum('pod_charge_memory_gigabyte_hours')
-                    },
-                    'capacity_aggregate': {
-                        'capacity': Max('cluster_capacity_memory_gigabyte_hours')
-                    },
-                    'default_ordering': {'usage': 'desc'},
-                    'annotations': {
-                        'usage': Sum('pod_usage_memory_gigabyte_hours'),
-                        'request': Sum('pod_request_memory_gigabyte_hours'),
-                        'limit': Sum('pod_limit_memory_gigabyte_hours'),
-                        'capacity': {
-                            'total': Max('total_capacity_memory_gigabyte_hours'),
-                            'cluster': Max('cluster_capacity_memory_gigabyte_hours'),
-                        },
-                        'infrastructure_cost': Value(0, output_field=DecimalField()),
-                        'derived_cost': Sum('pod_charge_memory_gigabyte_hours'),
-                        'cost': Sum('pod_charge_memory_gigabyte_hours'),
-                        'cost_units': Value('USD', output_field=CharField()),
-                        'usage_units': Value('GB-Hours', output_field=CharField())
-                    },
-                    'delta_key': {
-                        'usage': Sum('pod_usage_memory_gigabyte_hours'),
-                        'request': Sum('pod_request_memory_gigabyte_hours'),
-                        'cost': Sum('pod_charge_memory_gigabyte_hours')
-                    },
-                    'filter': {},
-                    'cost_units_key': 'USD',
-                    'usage_units_key': 'GB-Hours',
-                    'sum_columns': ['usage', 'request', 'limit', 'infrastructure_cost',
-                                    'derived_cost', 'cost'],
-                },
-                'volume': {
-                    'tables': {
-                        'query': OCPStorageLineItemDailySummary
-                    },
-                    'tag_column': 'volume_labels',
-                    'aggregates': {
-                        'usage': Sum('persistentvolumeclaim_usage_gigabyte_months'),
-                        'request': Sum('volume_request_storage_gigabyte_months'),
-                        'infrastructure_cost': Sum(Value(0, output_field=DecimalField())),
-                        'derived_cost': Sum('persistentvolumeclaim_charge_gb_month'),
-                        'cost': Sum('persistentvolumeclaim_charge_gb_month')
-                    },
-                    'capacity_aggregate': {
-                        'capacity': Sum('persistentvolumeclaim_capacity_gigabyte_months')
-                    },
-                    'default_ordering': {'usage': 'desc'},
-                    'annotations': {
-                        'usage': Sum('persistentvolumeclaim_usage_gigabyte_months'),
-                        'request': Sum('volume_request_storage_gigabyte_months'),
-                        'capacity': {
-                            'total': Sum('persistentvolumeclaim_capacity_gigabyte_months'),
-                            'cluster': Sum('persistentvolumeclaim_capacity_gigabyte_months'),
-                        },
-                        'infrastructure_cost': Value(0, output_field=DecimalField()),
-                        'derived_cost': Sum('persistentvolumeclaim_charge_gb_month'),
-                        'cost': Sum('persistentvolumeclaim_charge_gb_month'),
-                        'cost_units': Value('USD', output_field=CharField()),
-                        'usage_units': Value('GB-Mo', output_field=CharField()),
-                    },
-                    'delta_key': {
-                        'usage': Sum('persistentvolumeclaim_usage_gigabyte_months'),
-                        'request': Sum('volume_request_storage_gigabyte_months'),
-                        'cost': Sum('persistentvolumeclaim_charge_gb_month')
-                    },
-                    'filter': {},
-                    'cost_units_key': 'USD',
-                    'usage_units_key': 'GB-Mo',
-                    'sum_columns': ['usage', 'request', 'infrastructure_cost',
-                                    'derived_cost', 'cost'],
-                },
-            },
-            'start_date': 'usage_start',
-            'tables': {
-                'query': OCPUsageLineItemDailySummary,
-            },
-        },
-        {
-            'provider': 'OCP_AWS',
-            'alias': 'account_alias__account_alias',
-            'annotations': {
-                'cluster': 'cluster_id',
-                'project': 'namespace',
-                'account': 'usage_account_id',
-                'service': 'product_code',
-                'az': 'availability_zone'
-            },
-            'end_date': 'usage_end',
-            'filters': {
-                'project': {
-                    'field': 'namespace',
-                    'operation': 'icontains'
-                },
-                'cluster': [
-                    {
-                        'field': 'cluster_alias',
-                        'operation': 'icontains',
-                        'composition_key': 'cluster_filter'
-                    },
-                    {
-                        'field': 'cluster_id',
-                        'operation': 'icontains',
-                        'composition_key': 'cluster_filter'
-                    }
-                ],
-                'node': {
-                    'field': 'node',
-                    'operation': 'icontains'
-                },
-                'account': [
-                    {
-                        'field': 'account_alias__account_alias',
-                        'operation': 'icontains',
-                        'composition_key': 'account_filter'
-                    },
-                    {
-                        'field': 'usage_account_id',
-                        'operation': 'icontains',
-                        'composition_key': 'account_filter'
-                    }
-                ],
-                'service': {
-                    'field': 'product_code',
-                    'operation': 'icontains'
-                },
-                'product_family': {
-                    'field': 'product_family',
-                    'operation': 'icontains'
-                },
-                'az': {
-                    'field': 'availability_zone',
-                    'operation': 'icontains'
-                },
-                'region': {
-                    'field': 'region',
-                    'operation': 'icontains'
-                }
-            },
-            'group_by_options': ['account', 'service', 'region', 'cluster', 'project', 'node', 'product_family'],
-            'tag_column': 'tags',
-            'report_type': {
-                'costs': {
-                    'aggregates': {
-                        'infrastructure_cost': Sum(F('unblended_cost')),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum(F('unblended_cost')),
-                    },
-                    'annotations': {
-                        'infrastructure_cost': Sum(F('unblended_cost')),
-                        'derived_cost': Value(0, output_field=DecimalField()),
-                        'cost': Sum(F('unblended_cost')),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD'))
-                    },
-                    'count': None,
-                    'delta_key': {'cost': Sum(F('unblended_cost'))},
-                    'filter': {},
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'sum_columns': ['cost', 'infrastructure_cost', 'derived_cost'],
-                    'default_ordering': {'cost': 'desc'},
-                },
-                'costs_by_project': {
-                    'tables': {
-                        'query': OCPAWSCostLineItemProjectDailySummary,
-                        'total': OCPAWSCostLineItemProjectDailySummary
-                    },
-                    'aggregates': {
-                        'infrastructure_cost': Sum('pod_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('pod_cost'),
-                    },
-                    'annotations': {
-                        'infrastructure_cost': Sum('pod_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('pod_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD'))
-                    },
-                    'count': None,
-                    'delta_key': {'cost': Sum('pod_cost')},
-                    'filter': {},
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'sum_columns': ['infrastructure_cost',
-                                    'derived_cost', 'cost'],
-                    'default_ordering': {'cost': 'desc'},
-                },
-                'storage': {
-                    'aggregates': {
-                        'infrastructure_cost': Sum(F('unblended_cost')),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum(F('unblended_cost')),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'usage': Sum(F('usage_amount')),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'annotations': {
-                        'infrastructure_cost': Sum(F('unblended_cost')),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum(F('unblended_cost')),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'usage': Sum(F('usage_amount')),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'count': None,
-                    'delta_key': {'usage': Sum('usage_amount')},
-                    'filter': {
-                        'field': 'product_family',
-                        'operation': 'contains',
-                        'parameter': 'Storage'
-                    },
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'usage_units_key': 'unit',
-                    'usage_units_fallback': 'GB-Mo',
-                    'sum_columns': ['usage', 'infrastructure_cost',
-                                    'derived_cost', 'cost'],
-                    'default_ordering': {'usage': 'desc'},
-                },
-                'storage_by_project': {
-                    'tables': {
-                        'query': OCPAWSCostLineItemProjectDailySummary,
-                        'total': OCPAWSCostLineItemProjectDailySummary
-                    },
-                    'aggregates': {
-                        'infrastructure_cost': Sum('pod_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('pod_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'usage': Sum('usage_amount'),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'annotations': {
-                        'infrastructure_cost': Sum('pod_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('pod_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'usage': Sum('usage_amount'),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'count': None,
-                    'delta_key': {'usage': Sum('usage_amount')},
-                    'filter': {
-                        'field': 'product_family',
-                        'operation': 'contains',
-                        'parameter': 'Storage'
-                    },
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'usage_units_key': 'unit',
-                    'usage_units_fallback': 'GB-Mo',
-                    'sum_columns': ['usage', 'cost', 'infrastructure_cost', 'derived_cost'],
-                    'default_ordering': {'usage': 'desc'},
-                },
-                'instance_type': {
-                    'aggregates': {
-                        'infrastructure_cost': Sum(F('unblended_cost')),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum(F('unblended_cost')),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'count': Count('resource_id', distinct=True),
-                        'usage': Sum(F('usage_amount')),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'aggregate_key': 'usage_amount',
-                    'annotations': {
-                        'infrastructure_cost': Sum(F('unblended_cost')),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum(F('unblended_cost')),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'count': Count('resource_id', distinct=True),
-                        'count_units': Value('instances', output_field=CharField()),
-                        'usage': Sum(F('usage_amount')),
-                        'usage_units': Coalesce(Max('unit'), Value('Hrs'))
-                    },
-                    'count': 'resource_id',
-                    'delta_key': {'usage': Sum('usage_amount')},
-                    'filter': {
-                        'field': 'instance_type',
-                        'operation': 'isnull',
-                        'parameter': False
-                    },
-                    'group_by': ['instance_type'],
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'usage_units_key': 'unit',
-                    'usage_units_fallback': 'Hrs',
-                    'count_units_fallback': 'instances',
-                    'sum_columns': ['usage', 'cost', 'infrastructure_cost', 'derived_cost', 'count'],
-                    'default_ordering': {'usage': 'desc'},
-                },
-                'instance_type_by_project': {
-                    'tables': {
-                        'query': OCPAWSCostLineItemProjectDailySummary,
-                        'total': OCPAWSCostLineItemProjectDailySummary
-                    },
-                    'aggregates': {
-                        'infrastructure_cost': Sum('pod_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('pod_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'count': Count('resource_id', distinct=True),
-                        'usage': Sum('usage_amount'),
-                        'usage_units': Coalesce(Max('unit'), Value('GB-Mo'))
-                    },
-                    'aggregate_key': 'usage_amount',
-                    'annotations': {
-                        'infrastructure_cost': Sum('pod_cost'),
-                        'derived_cost': Sum(Value(0, output_field=DecimalField())),
-                        'cost': Sum('pod_cost'),
-                        'cost_units': Coalesce(Max('currency_code'), Value('USD')),
-                        'count': Count('resource_id', distinct=True),
-                        'count_units': Value('instances', output_field=CharField()),
-                        'usage': Sum('usage_amount'),
-                        'usage_units': Coalesce(Max('unit'), Value('Hrs'))
-                    },
-                    'count': 'resource_id',
-                    'delta_key': {'usage': Sum('usage_amount')},
-                    'filter': {
-                        'field': 'instance_type',
-                        'operation': 'isnull',
-                        'parameter': False
-                    },
-                    'group_by': ['instance_type'],
-                    'cost_units_key': 'currency_code',
-                    'cost_units_fallback': 'USD',
-                    'usage_units_key': 'unit',
-                    'usage_units_fallback': 'Hrs',
-                    'count_units_fallback': 'instances',
-                    'sum_columns': ['usage', 'cost', 'infrastructure_cost', 'derived_cost', 'count'],
-                    'default_ordering': {'usage': 'desc'},
-                },
-            },
-            'start_date': 'usage_start',
-            'tables': {
-                'query': OCPAWSCostLineItemDailySummary,
-                'total': OCPAWSCostLineItemDailySummary
-            },
-        }
-    ]
-
-    @staticmethod
-    def provider_data(provider):
-        """Return provider portion of map structure."""
-        for item in ProviderMap.MAPPING:
-            if provider in item.get('provider'):
-                return item
-
-    @staticmethod
-    def report_type_data(report_type, provider):
-        """Return report_type portion of map structure."""
-        prov = ProviderMap.provider_data(provider)
-        return prov.get('report_type').get(report_type)
-
-    def __init__(self, provider, report_type):
-        """Constructor."""
-        self._provider = provider
-        self._report_type = report_type
-        self._provider_map = ProviderMap.provider_data(provider)
-        self._report_type_map = ProviderMap.report_type_data(report_type, provider)
-
-    @property
-    def count(self):
-        """Return the count property."""
-        return self._report_type_map.get('count')
-
-    @property
-    def provider_map(self):
-        """Return the provider map property."""
-        return self._provider_map
-
-    @property
-    def query_table(self):
-        """Return the appropriate query table for the report type."""
-        report_table = self._report_type_map.get('tables', {}).get('query')
-        default = self._provider_map.get('tables').get('query')
-        return report_table if report_table else default
-
-    @property
-    def report_type_map(self):
-        """Return the report-type map property."""
-        return self._report_type_map
-
-    @property
-    def sum_columns(self):
-        """Return the sum column list for the report type."""
-        return self._report_type_map.get('sum_columns')
-
-    @property
-    def tag_column(self):
-        """Return the appropriate query table for the report type."""
-        report_specific_column = self._report_type_map.get('tag_column')
-        default = self._provider_map.get('tag_column')
-        return report_specific_column if report_specific_column else default
-
-    @property
-    def cost_units_key(self):
-        """Return the cost_units_key property."""
-        return self._report_type_map.get('cost_units_key')
-
-    @property
-    def usage_units_key(self):
-        """Return the usage_units_key property."""
-        return self._report_type_map.get('usage_units_key')
-
-
 class ReportQueryHandler(QueryHandler):
     """Handles report queries and responses."""
 
-    def __init__(self, query_parameters, url_data,
-                 tenant, **kwargs):
+    def __init__(self, query_parameters, tenant, **kwargs):
         """Establish report query handler.
 
         Args:
@@ -811,33 +53,46 @@ class ReportQueryHandler(QueryHandler):
         LOG.debug(f'Query Params: {query_parameters}')
         self._accept_type = None
         self._group_by = None
-        self._tag_keys = []
         self._access = {}
-        if kwargs:
-            # view parameters
-            elements = ['accept_type', 'delta', 'report_type', 'tag_keys', 'access']
-            for key, value in kwargs.items():
-                if key in elements:
-                    setattr(self, f'_{key}', value)
 
         assert getattr(self, '_report_type'), \
             'kwargs["report_type"] is missing!'
-        self._mapper = ProviderMap(provider=kwargs.get('provider'),
-                                   report_type=self._report_type)
         default_ordering = self._mapper._report_type_map.get('default_ordering')
 
-        super().__init__(query_parameters, url_data,
-                         tenant, default_ordering, **kwargs)
-
-        self.group_by_options = self._mapper._provider_map.get('group_by_options')
+        super().__init__(query_parameters, tenant, default_ordering, **kwargs)
 
         self._delta = self.query_parameters.get('delta')
-        self._limit = self.get_query_param_data('filter', 'limit')
         self._offset = self.get_query_param_data('filter', 'offset', default=0)
         self.query_delta = {'value': None, 'percent': None}
 
         self.query_filter = self._get_filter()
         self.query_exclusions = self._get_exclusions()
+
+    def _initialize_kwargs(self, kwargs):
+        """Initialize class attributes from view kwargs.
+
+        Params:
+            kwargs (dict) - kwargs dict received from __init__()
+        """
+        # view parameters
+        elements = ['accept_type', 'delta', 'report_type', 'tag_keys', 'access']
+        for key, value in kwargs.items():
+            if key in elements:
+                # don't overwrite existing attributes.
+                try:
+                    getattr(self, f'_{key}')
+                except AttributeError:
+                    setattr(self, f'_{key}', value)
+
+        for key in elements:
+            # set defaults for required attributes.
+            try:
+                getattr(self, f'_{key}')
+            except AttributeError:
+                if key == 'tag_keys':
+                    setattr(self, f'_{key}', [])
+                else:
+                    setattr(self, f'_{key}', None)
 
     def initialize_totals(self):
         """Initialize the total response column values."""
